# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Agent Runtime 示例应用 API 文档。",
        "title": "Agent Runtime API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/workspaces/users/{user_id}": {
            "get": {
                "description": "管理员查看指定用户 home/task workspace 状态，并写入后台审计记录。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-workspaces"
                ],
                "summary": "查看用户工作区摘要",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户 ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.UserWorkspaceSummarySwaggerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorSwaggerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorSwaggerResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorSwaggerResponse"
                        }
                    }
                }
            }
        },
...
        "/skills": {
            "get": {
                "description": "返回当前 workspace 下全部非 hidden 的技能列表，不包含 content。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skills"
                ],
                "summary": "获取技能列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SkillListSwaggerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorSwaggerResponse"
                        }
                    }
                }
            }
        },
        "/skills/{name}": {
            "get": {
                "description": "根据技能名返回 workspace skill 详情，允许读取 hidden 技能。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skills"
                ],
                "summary": "获取技能详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "技能名",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SkillSwaggerResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorSwaggerResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ErrorSwaggerResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                },
                "time": {
                    "type": "string"
                }
            }
        },
...
        "handlers.SkillListItemDoc": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "hidden": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "source_ref": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tools": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "handlers.SkillListSwaggerResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.SkillListItemDoc"
                    }
                },
                "message": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                },
                "time": {
                    "type": "string"
                }
            }
        },
        "handlers.SkillSwaggerDoc": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "hidden": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "resource_refs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "source_ref": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tools": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "handlers.SkillSwaggerResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/handlers.SkillSwaggerDoc"
                },
                "message": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                },
                "time": {
                    "type": "string"
                }
            }
        },
...
        "handlers.TaskWorkspaceSummarySwaggerDoc": {
            "type": "object",
            "properties": {
                "backup_root": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "discarded_at": {
                    "type": "string"
                },
                "merged_at": {
                    "type": "string"
                },
                "mode": {
                    "type": "string",
                    "enum": [
                        "mutable",
                        "readonly"
                    ]
                },
                "state": {
                    "type": "string",
                    "enum": [
                        "active",
                        "pending_merge",
                        "merged",
                        "discarded",
                        "completed"
                    ]
                },
                "task_id": {
                    "type": "string"
                },
                "task_root": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "handlers.UserWorkspaceSummarySwaggerDoc": {
            "type": "object",
            "properties": {
                "home_root": {
                    "type": "string"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.TaskWorkspaceSummarySwaggerDoc"
                    }
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "handlers.UserWorkspaceSummarySwaggerResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/handlers.UserWorkspaceSummarySwaggerDoc"
                },
                "message": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                },
                "time": {
                    "type": "string"
                }
            }
        },
...
    }
}